Moving all map stuff to one spot
