+++ server/map/mapLoader.ts
@@ -1,9 +1,11 @@
 import tmx from "tmx-parser";
 import path from "path";
 
+const MAP_DIR = path.join(__dirname, "map-files");
+
 export async function loadMap() {
   const map = await new Promise((resolve, reject) => {
-    tmx.parseFile(path.join(__dirname, "map.tmx"), function (err, loadedMap) {
+    tmx.parseFile(path.join(MAP_DIR, "map.tmx"), function (err, loadedMap) {
       if (err) return reject(err);
       resolve(loadedMap);
     });
